[owners] Remove paulsemel@chromium.org from testing/libfuzzer/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/paulsemel@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I10a48ebcc1e24e7fdf38dd92e85256af069f4ae0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6931470
Reviewed-by: Pâris Meuleman <pmeuleman@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Pâris Meuleman <pmeuleman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1514159}
NOKEYCHECK=True
GitOrigin-RevId: 35999cd3f9993465987de08a1f0f14136660e5f0