[lensoverlay] Remove start query flow optimization feature.
This is a partial cleanup of the `LensOverlayLatencyOptimizations`
feature. This is being done to remove complexity from the implementation
to reinvoke the live page.

This also removes `LensOverlayLatencyOptimizations` from the field trial
configuration, as well as other expired studies related to the Lens
overlay.

Bug: 445698141
Change-Id: I4695c0bf897e39e9204051763885ea4e3379fb14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6963043
Reviewed-by: Jason Hu <hujasonx@google.com>
Commit-Queue: Juan Mojica <juanmojica@google.com>
Cr-Commit-Position: refs/heads/main@{#1517015}
NOKEYCHECK=True
GitOrigin-RevId: 0c169eaa614d89b53964c25ef40917f9b6b72d7e