Add Autofill observer to trigger a CSD ping on each credit card form.
This updates ClientSideDetectionHost to self register as an
AutofillManager::Observer, with the OnFieldTypesDetermined callback
implemented. It checks for the form type kCreditCardForm to trigger
sending a CSD ping.

Behavior is gated behind the ClientSideDetectionAutofillCreditCardForm
feature (disabled by default). Additionally, it has two parameters.
HCAcceptanceRate defaults to 1.0 and can be set to a lower number so
that some percentage of events that otherwise match the high-confidence
allowlist instead proceed to pre-classification. SampleRate defaults to
0.0 and will be set to a higher number that specifies what percentage of
events should produce a CSD ping.

The internal repo change to csd.proto is in cl/803130110.

Bug: 443098659
Change-Id: Ie89ffe35c0ae0983bb8f0dc5e934b291b32fb4d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6911850
Commit-Queue: Brian Begnoche <bcb@chromium.org>
Reviewed-by: Nwokedi Idika <nwokedi@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1517874}
NOKEYCHECK=True
GitOrigin-RevId: 229331b05e18726a2f2e4f5fab2d0bfcc15145eb