[Browser Frame] Enable new layout logic in test config
Make test builds default to using the new browser layout instead of the
old one via fieldtrial_testing_config.json.

This also allows us to roll the change into Canary.

In order to make all tests pass, whether or not a web app toolbar is
displayed needed to be consistent between old and new implementation.
In the old API, whether or not a toolbar should be shown is buried in
`GetBoundsForWebAppFrameToolbar()`, which returns an empty bounds if
the toolbar shouldn't show. However, the new API does not use this
method. In this CL, the logic for whether a web app toolbar should be
shown is broken out into its own method, and used by both the old and
new codepaths.

A few other fixes are included for layout calculation:
 - On Mac, now correctly take into account the fact that the caption
   (i.e. "stoplight") buttons are not part of the main browser window
   when in fullscreen/immersive mode.
 - On Windows, PWAs with custom overlays now correctly take into
   account that the window icon is hidden when the overlay is active.

Include-Ci-Only-Tests: chromium.mac:mac15-x64-rel-tests|browser_tests
Bug: 443122496
Change-Id: If164b3ce75bca907baca4d3c4883c05a6e8f480a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6990550
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1524222}
NOKEYCHECK=True
GitOrigin-RevId: 8c674a24923cedeb28ac630793309fbfd0ff666f