Add screenshot testing for Lens composebox
This CL adds the base for screenshotting Lens side panel elements,
starting with the Lens composebox. The test is setup in a way to be
easily expandable to other elements, sharing common setup in a base
class.

While writing these screenshots, I discovered to issues in the testing
infra that I needed to fix:
1) Unable to load chrome-unstrusted:// URLs in a tab. To fix this, I
   had to add chrome-untrusted to the allowlist that already allowed
   chrome:// URLs by allowing generalized WebUI URLs.
2) Focus being cleared prior to the screenshot being taken. This is a
   measure in place to prevent flakiness, but makes it impossible to
   test JS focus behavior. Instead of mocking the focus logic, adds a
   struct to pass options to the screenshotting to allow clearing focus
   to be skipped for WebUI tests.

Change-Id: If49a950304177c560845d8845c499cc32cf72ed6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7018205
Reviewed-by: Christopher Grant <cjgrant@chromium.org>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1531503}
NOKEYCHECK=True
GitOrigin-RevId: 756a260ddb2ae430bf90eb22c12528932ac1b4c8