Reland: Add finch field trial for chromevox in mv3
This is a reland of https://crrev.com/c/6962374. The original change
was reverted due to failing msan tests, which we are disabling for now
to unblock the finch rollout. Small adjustments have been made to two
test suites to ensure they pass if ChromeVox in mv3 is enabled (which
happens for a small number of builders when the finch field trial
is added).

Bug: 433771715
Change-Id: Ic893558d472b5aa50039be437afdcebf2accd0d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7047197
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531639}
NOKEYCHECK=True
GitOrigin-RevId: f8823534986814cc0cbeecf0c225c214c3f5eaf7