Add LRU cache to gfx:Canvas::GetStringWidth()/GetStringWidthF()
This CL routes GetStringWidth() to use the result of GetStringWidthF(),
and adds caching to GetStringWidthF() for all strings with a length of
at most 4.

This was decided after observing that Label::GetMinimumSize() calls
gfx:Canvas::GetStringWidth() repeatedly with an ellipsis character and
the same font_list. Further analysis showed other strings that are
repeatedly measured. It is part of a broader effort to reduce the weight
of unnecessary operations.

PlatformFont is used as the key for the cache (together with the text
itself), as it includes the UniqueID from the SkTypeface, the font size,
and the font style.

Bug: 448407151
Change-Id: I5c53705a87ad4dd27ba06cd1e763c3b787d91e41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6999431
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1531751}
NOKEYCHECK=True
GitOrigin-RevId: 2c779ea3f3c974e28e344491bb189358f259ff35