[Predictable App Updating] Add field trial testing configs
This CL adds field trial testing configs for ensuring that there are
no test failures when the predictable app updating feature is
enabled. The current failures are in the manifest update manager
browser tests and web app integration tests only. The former is run
with the feature turned off since those are tightly coupled to the
older implementation of manifest updating. The integration test
failures will be handled in a follow-up.

Bug: 452009643
Change-Id: Ic3b028cf3b0c7fd80f2efd24b5dd160e244f6346
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7038748
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532328}
NOKEYCHECK=True
GitOrigin-RevId: d41404d2b054402bca9589314ae5b78250865f96