Reland: Add finch field trial for chromevox in mv3
The original attempt [1] was reverted due to flaky tests on msan
builds. To unblock this change, this change forces the flaky test
suite to run ChromeVox in manifest v2. We will get the tests working
in manifest v3 in a follow-up CL.

[1] https://crrev.com/c/7047197

Bug: 433771715
Change-Id: Ia84e23d7aa3ed220d8ee54c4582d8c8099d443b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7063014
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1532718}
NOKEYCHECK=True
GitOrigin-RevId: 0cb46e171117604f1334b15adb420fb0cfd3148c