[iOS] Add DriveFileUploader::FetchSaveToDriveClientFolder() method
This CL adds a `FetchSaveToDriveClientFolder()` API to the Drive file
uploader class which can be used to perform a get-or-create operation in
the user's Drive in a single API call, instead of the doing the current
search-then-create operations which may require up to two calls, if the
Save to Drive destination folder does not exist and needs to be created.

Also DriveUploadTask now uses this new API instead of search-then-create
if the kIOSSaveToDriveClientFolder feature flag is enabled.

Bug: 452543744
Change-Id: I271203e574ce5ff411bad4eca6aa7d2e8e499495
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7063593
Commit-Queue: Quentin Pubert <qpubert@google.com>
Auto-Submit: Quentin Pubert <qpubert@google.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1532872}
NOKEYCHECK=True
GitOrigin-RevId: 71c0d1a85a948c648bd4e01ed3d628fa43d1588b