Experimental feature to allow the navigation to resume after speculative process kills
The CL adds a feature to allow the navigation before committing to
resume event if the process for the speculative RFH is killed. A new RFH
will be automatically created in the next call to
GetFrameHostForNavigation. The CL also adds a UMA for collecting the
frequency of such events and whether it is possible to recover.

Bug: 453878130
Change-Id: I6192de2f1ee49a801fe2bcbe283b245f92668e19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7062071
Commit-Queue: Jiacheng Guo <gjc@google.com>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1534784}
NOKEYCHECK=True
GitOrigin-RevId: 722ee00276787298bac7a4a109f909c32f12164d