Only drop saved EventMetrics for FrameSkippedReason::kNoDamage
This CL modifies cc::LayerTreeHost::DidNotProduceFrame() to only drop
saved event metrics if the associated frame wasn't produced because
there was no damage.

In all other cases (e.g. if the frame wasn't produced because it was
waiting on main), the saved event metrics will be preserved and
associated with the next presented frame. In practice, this change might
cause small regressions in latency and/or jank metrics. Note that this
wouldn't be a user-visible change but a correction of the metrics.

The change is guarded behind a new feature,
"DropMetricsFromNonProducedFramesOnlyIfTheyHadNoDamage".

Bug: b:444183591
Change-Id: I277ecc0ac57e93a225771597a48413c5f805d7a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7124401
Auto-Submit: Petr Cermak <petrcermak@chromium.org>
Reviewed-by: Kartar Singh <kartarsingh@google.com>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Petr Cermak <petrcermak@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1541235}
NOKEYCHECK=True
GitOrigin-RevId: ce33b36e035c37d5212df48045dbf01cf47468aa