Enable AutofillAndPasswordsInSameSurface
This CL enables the AutofillAndPasswordsInSameSurface by default.

Some browser tests still remain that check expectations for the old call
path into the PasswordAutofillAgent and those test will fail with the
feature enabled. To mitigate, this CL disables the feature for those
tests and immediate followup work is needed to change the expectations
to check for the delegation calls into to Autofill infra.

The current alteration to the tests adds a new ScopedFeatureList to each
affected test case. Since this is meant to be a short-term temporal
change, the existing ScopedFeatureList member of the fixture is not
extended to also cover this flag.

Change-Id: Ic82c355fc01cd7cd6702beb00f5765db4216700f
Bug: 445402617, 460655995, 460645840
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7151338
Commit-Queue: Matthias Körber <koerber@google.com>
Reviewed-by: Dominic Battre <battre@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Auto-Submit: Matthias Körber <koerber@google.com>
Cr-Commit-Position: refs/heads/main@{#1544814}
NOKEYCHECK=True
GitOrigin-RevId: db054f382f000f7368827472152d584c2ea191af