Migrate Lens Homework chip to PageActionController
This change migrates the Lens Overlay Homework page action to the modern
PageActionController framework. This is part of the ongoing effort to
unify all page action implementations under a single, more maintainable
system.

The core of this change is the introduction of the new
`LensOverlayHomeworkPageActionController`. This controller now
encapsulates all the logic for determining when the homework action chip
should be displayed to the user.

A new feature parameter, `kPageActionsMigrationLensOverlayHomework`, has
been added to the `kPageActionsMigration` feature flag to control this
migration. The previous implementation path is preserved and will be
used when this parameter is disabled.

Bug: 444416267
Change-Id: If51729b6d8e5912ef4433f7fe7149884e8d3fcf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7128255
Reviewed-by: Darryl James <dljames@chromium.org>
Reviewed-by: Kaan Alsan <alsan@chromium.org>
Commit-Queue: Daniel Soromou <koretadaniel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1545318}
NOKEYCHECK=True
GitOrigin-RevId: 52ade4a309763e3dbe84d6987dd290bf1a3340b0