[TreesInViz] Turn on the feature for Linux in fieldtrial.
This way, we have testing coverage for both TreesInViz enabled
and disabled.

Due to the nature of this feature, so far we only encountered
bugs related to TreesInViz either specific to Desktop platforms,
or to Android, but never to a specific desktop platform.

TEST=bots

Bug: 369883417
Change-Id: Ic59abca900487dc5bb3f7e8335bd9411aa903dda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7152659
Reviewed-by: vikas soni <vikassoni@chromium.org>
Commit-Queue: Zhenyao Mo <zmo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1545969}
NOKEYCHECK=True
GitOrigin-RevId: 4bf73a4bcf98bc986a921a63af23d25282dbccb2