LoadAllTabsAtStartup testing config
LoadAllTabsAtStartup is Android only, but LazyBrowserInterfaceBroker
can be rolled out on all platforms as it is largely just a performance
optimization.

For Android the two flags will use a coupled rollout since we want to
measure the impact of loading all tabs only if the memory optimization
is enabled otherwise the memory regressions will be even worse.

Bug: 450912216
Change-Id: I79df3682ae5f269bc91054fefb54d82669ecd29b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7127302
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547147}
NOKEYCHECK=True
GitOrigin-RevId: 3ae9f8a9a11162ce4000f44aac8547db946020f4