[a11y] Only request text cursor on screen for caret moves
The initial implementation of Magnification-follows-text-cursor
(crrev.com/c/6986642) requested that the caret be placed on-screen on
every frame update from the renderer, preventing scrolling when a text
caret was active. This moves the view request to a path that is only
called when the caret moves, fixing that bug.

Given the bug in the initial launch, this re-launch puts the feature
behind a Finch flag, and, given the availability of a Finch kill-switch,
enables the feature for WebView as well and also the non-source-tagged
`View.requestRectangleOnScreen(Rect)` fallback API for pre-36.1 Android
API levels.

AX-Relnotes: Adds Magnification-follows-text-cursor feature for all supported Android versions, guarded behind `AccessibilityMagnificationFollowsFocus` flag.
Bug: 454771436, 443079767
Change-Id: Ib894b95ca69e623f8fde7703431f125342f9e97e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7128421
Commit-Queue: Aaron Moss <aaronmoss@google.com>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1549241}
NOKEYCHECK=True
GitOrigin-RevId: 52bf4880a7c20f8e954d8fc6ccce4c185e7fcbac