Add module scheme to test binaries
Module scheme is used by structured test ids to enable the results to be uploaded in the new format.

Adds it to model_validation_tests and ondevice_stability_tests which are ran on internal bots.

This CL needs to land first:
  https://chrome-internal-review.googlesource.com/c/chrome/components/optimization_guide/+/8812098

Bug: 440384377
Change-Id: I48979aa9d9d170fd74b623c2237d13dbeff6365f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7218815
Commit-Queue: Benjamin Joyce (Ben) <bjoyce@google.com>
Reviewed-by: Haiyang Pan <hypan@google.com>
Cr-Commit-Position: refs/heads/main@{#1553451}
NOKEYCHECK=True
GitOrigin-RevId: 644d3b749a0a7cafb676a4aba01879bc8fec077c