Add a feature param to initialize feature list early for Android
This CL adds a new feature param called InitFeatureListEarly.
When it is enabled, we will initialize feature list early, immediately
after loading the native library. In order to do that, we will need to
initialize a number of things early, such as creating the thread pool
and creating the variations ids provider.

[Gardeners]: Please don't revert this CL for simple test failures,
since this CL is a large-scale change and a few simple test failures
are expected.

A doc with more explanations of the project can be found here:
https://docs.google.com/document/d/1UNx_fgFjGcp9bjQEMfP_K4Ly1f7-JCO4SNHJN4oBF20

Bug: 455890560
Change-Id: I8bb9703e8a19e0528a77ed76d9e57bae8b8b0cad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7092856
Commit-Queue: Martin Kong <martinkong@google.com>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Owners-Override: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1554114}
NOKEYCHECK=True
GitOrigin-RevId: 40ed9989f483360aa4e4ed9f7f4605f201853bc2