Rearrange the XSLT feature flags
There was a bug where the special trial flag being disabled meant that
the about:flags XSLT setting could not be used to re-enable XSLT. That's
bad, because we want developers to be able to test XSLT
enabled/disabled, and also use pre-stable Chrome.

This CL reorganizes the flags, and centralizes checking into one static
method in XSLTProcessor. This also puts back the single runtime enabled
feature flag, which allows the IDL to key off of it, and remove the
XSLTProcessor API entirely when the flag is disabled. That was also
causing another issue where just feature-detecting XSLT caused the
big deprecation warning message.

Note that because of the change in definitions, this will require a
restart of the XSLT Finch experiment.

Bug: 435623334
Change-Id: I624cd9e2ba9c57f531ce52b4556b53133c9a3b72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7229917
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1554870}
NOKEYCHECK=True
GitOrigin-RevId: 13f370ad2ed32e0cec0d33847fd832613294fd7b