[TreesInViz] Change tiling removal behavior.
Currently tiling removal is triggered by CleanUpTilingsOnActiveLayer,
which is partly from AppendQuads. However, AppendQuads isn't called in
client side.

In this CL, we add CleanUpTilingsOnActiveLayer to PictureLayerImpl's
DidDraw.

We can remove it from AppendQuads now and align TreesInViz and non
TreesInviz behaviors. However, this leads to a bunch of test failures,
so we will do that in a separate followup CL.

This CL also fixes LayerTreeHostTestCrispUpAfterPinchEnds test by
inserting a frame to wait for the tiling to be removed.

TEST=cc_unittests

Bug: 448683984
Change-Id: Idbb6e3a9d7c58cda350acfdd7e7cfd250ec49bf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7277489
Commit-Queue: Zhenyao Mo <zmo@chromium.org>
Auto-Submit: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Vikas Soni <vikassoni@chromium.org>
Reviewed-by: Victor Miura <vmiura@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1561367}
NOKEYCHECK=True
GitOrigin-RevId: dfb14790598cac2c20058ca4bf9eb9bd9af78bab